move checkpointing utils
